styled components/sass cleanups
