+++ scss/_variables.scss
@@ -45,7 +45,7 @@ $color-message-detail-delete-button-border: #a4a6a9;
 $color-message-detail-delete-bg: #f44336;
 $color-message-detail-delete: #ffffff;
 
-$color-conversation-list-item-default-avatar-bg: #505050; //This is never used because the core allways sets the color
+$color-chat-list-item-default-avatar-bg: #505050; //This is never used because the core allways sets the color
 $color-global-textarea-outline: #2090ea;
 
 // A few layout variables used cross-file
