+++ scss/modules/_chat-list-item.scss
@@ -1,28 +1,30 @@
 /* Chat List (conversation list) */
-$color-conversation-list-has-unread: var(--unreadCountBg);
-$color-conversation-list-unread-count-bg: var(--unreadCountBg);
-$color-conversation-list-unread-count-label: var(--unreadCountLabel);
-$color-conversation-list-message-error: var(--errorColor);
-$color-conversation-list-avatar-text: var(--avatarLabelColor);
+$color-chat-list-item-has-unread: var(--unreadCountBg);
+$color-chat-list-item-unread-count-bg: var(--unreadCountBg);
+$color-chat-list-item-unread-count-label: var(--unreadCountLabel);
+$color-chat-list-item-message-error: var(--errorColor);
+$color-chat-list-item-avatar-text: var(--avatarLabelColor);
 $chatlist-isgroup-icon: var(--globalText);
 
-.module-conversation-list-item {
+.chat-list-item {
   display: flex;
   flex-direction: row;
-  padding-right: 16px;
-  padding-left: 16px;
+  padding-right: 10px;
+  padding-left: 10px;
   align-items: center;
 
   cursor: pointer;
-
+  &:hover {
+    background-color:#bbb;
+  }
 }
 
-.module-conversation-list-item--has-unread {
+.chat-list-item--has-unread {
   padding-left: 12px;
-  border-left: 4px solid $color-conversation-list-has-unread;
+  border-left: 4px solid $color-chat-list-item-has-unread;
 }
 
-.module-conversation-list-item__avatar {
+.chat-list-item__avatar {
   margin-top: 8px;
   margin-bottom: 8px;
   height: 48px;
@@ -32,16 +34,16 @@ $chatlist-isgroup-icon: var(--globalText);
 
   object-fit: cover;
 }
-.module-conversation-list-item__default-avatar {
-  color: $color-conversation-list-avatar-text;
+.chat-list-item__default-avatar {
+  color: $color-chat-list-item-avatar-text;
   font-size: 26px;
   line-height: 48px;
   text-align: center;
 
-  background-color: $color-conversation-list-item-default-avatar-bg;
+  background-color: $color-chat-list-item-default-avatar-bg;
 }
 
-.module-conversation-list-item__content {
+.chat-list-item__content {
   flex-grow: 1;
   margin-left: 12px;
   // parent - 48px (for avatar) - 16px (our right margin)
@@ -52,13 +54,13 @@ $chatlist-isgroup-icon: var(--globalText);
   align-items: stretch;
 }
 
-.module-conversation-list-item__header {
+.chat-list-item__header {
   display: flex;
   flex-direction: row;
   align-items: center;
 }
 
-.module-conversation-list-item__header__name {
+.chat-list-item__header__name {
   flex-grow: 1;
   flex-shrink: 1;
   font-size: 14px;
@@ -75,11 +77,11 @@ $chatlist-isgroup-icon: var(--globalText);
   }
 }
 
-.module-conversation-list-item__header__name--with-unread {
+.chat-list-item__header__name--with-unread {
   font-weight: 300;
 }
 
-.module-conversation-list-item__header__timestamp {
+.chat-list-item__header__timestamp {
   flex-shrink: 0;
   margin-left: 6px;
 
@@ -94,11 +96,11 @@ $chatlist-isgroup-icon: var(--globalText);
   text-transform: uppercase;
 }
 
-.module-conversation-list-item__header__date--has-unread {
+.chat-list-item__header__date--has-unread {
   font-weight: 300;
 }
 
-.module-conversation-list-item__message {
+.chat-list-item__message {
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -106,7 +108,7 @@ $chatlist-isgroup-icon: var(--globalText);
   margin-top: 3px;
 }
 
-.module-conversation-list-item__message__text {
+.chat-list-item__message__text {
   flex-grow: 1;
   flex-shrink: 1;
 
@@ -119,6 +121,18 @@ $chatlist-isgroup-icon: var(--globalText);
   text-overflow: ellipsis;
 }
 
-.module-conversation-list-item__message__text--has-unread {
+.chat-list-item__message__text--has-unread {
   font-weight: 300;
 }
+
+.chat-list-item__is-verified {
+  width: 0.75em;
+  height: 0.75em;
+  margin-right: 2px;
+}
+
+.forward-message-list-chat-list .chat-list-item {
+    padding-right: 20px;
+    padding-left: 20px;
+}
+
