+++ scss/modules/_contact-list.scss
@@ -41,8 +41,3 @@ $color-contact-list-item: var(--contactListItemBg);
   margin-bottom: -1px;
 }
 
-.module-conversation-list-item__is-verified {
-  width: 0.75em;
-  height: 0.75em;
-  margin-right: 2px;
-}
