+++ scss/modules/_forward-message.scss
@@ -1,0 +1,4 @@
+.forward-message-list-chat-list .chat-list-item {
+    padding-right: 20px;
+    padding-left: 20px;
+}
