+++ scss/modules/index.scss
@@ -1,6 +1,8 @@
 @import 'contact-list';
 @import 'context-menu';
-@import 'conversation-list';
+@import 'chat-list';
+@import 'chat-list-item';
+@import 'forward-message';
 @import 'map';
 @import 'message';
 @import 'message-detail';
